refactor(questionnaires): split questionnaires/[id] page under the line cap (#544)
Extract the edit page's Event Assignments card into
`QuestionnaireEditAssignments` (owning its `getEventTypeLabel` /
`getSeriesEventCount` defensive readers + `formatDate`) and the Questions &
Sections card into `QuestionnaireEditQuestionsCard`. The questions card takes
the shared `QuestionnaireBuilder` instance plus `canEdit` / `showLlmGuidelines`
/ `questionsError`; calling `builder.method()` keeps the `this` binding and the
class `$state` stays reactive across the component boundary. Read-only fallback,
canEdit gating and the QuestionEditor/SectionEditor keying are unchanged.

`[id]/+page.svelte`: 872 -> 464 lines.
